// Qanat.Web/src/app/shared/generated/model/custom-rich-text-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CustomRichTextTypeSimpleDto } from '././custom-rich-text-type-simple-dto';
import { GeographySimpleDto } from '././geography-simple-dto';

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class CustomRichTextDto { 
    CustomRichTextID?: number;
    CustomRichTextType?: CustomRichTextTypeSimpleDto;
    CustomRichTextTitle?: string;
    CustomRichTextContent?: string;
    Geography?: GeographySimpleDto;
    IsEmptyContent?: boolean;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface CustomRichTextDtoForm { 
    CustomRichTextID?: FormControl<number>;
    CustomRichTextType?: FormControl<CustomRichTextTypeSimpleDto>;
    CustomRichTextTitle?: FormControl<string>;
    CustomRichTextContent?: FormControl<string>;
    Geography?: FormControl<GeographySimpleDto>;
    IsEmptyContent?: FormControl<boolean>;
}

export class CustomRichTextDtoFormControls { 
    public static CustomRichTextID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomRichTextType = (value: FormControlState<CustomRichTextTypeSimpleDto> | CustomRichTextTypeSimpleDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<CustomRichTextTypeSimpleDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomRichTextTitle = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CustomRichTextContent = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Geography = (value: FormControlState<GeographySimpleDto> | GeographySimpleDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<GeographySimpleDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsEmptyContent = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}